Let `compose` support `setup.steps`
